part 2 - join rows modules https://online.sbis.ru/opendoc.html?guid=d26eb342-c3ff-4892-84af-27819a43196a
