+++ Controls/_treeGrid/render/NodeFooterTemplate.wml
@@ -1,7 +1,7 @@
 <div class="{{ (gridColumn || itemData).getContentClasses({
     'hasContent': !!content || !!contentTemplate
 }) }}">
-    <ws:if data="{{ item.shouldDisplayVisibleFooter(!!content || !!contentTemplate) }}">
+    <ws:if data="{{ item.shouldDisplayExtraItem(!!content || !!contentTemplate) }}">
         <ws:partial if="{{ beforeContentTemplate }}" template="{{ beforeContentTemplate }}"/>
 
         <ws:if data="{{ gridColumn.shouldRenderHasMoreButton() }}">
@@ -11,7 +11,7 @@
                         loadMoreCaption="{{ item.getMoreCaption() }}"
                         linkFontColorStyle="{{ item.getMoreFontColorStyle() }}"
                         linkFontSize="xs"
-                        linkClass="controls-TreeGrid__nodeFooterLoadMore"/>
+                        linkClass="{{ item.getMoreClasses() }}"/>
         </ws:if>
         <ws:else data="{{ !gridColumn.isMoreButtonFooter() }}">
             <div class="controls-TreeGrid__nodeFooter-cell__customContent">
