+++ Controls/_treeGrid/styles/_TreeGridView.less
@@ -84,8 +84,7 @@
    min-width: @_expander_width_treeGrid;
 }
 
-.controls-TreeGrid__node-footer-expanderPadding,
-.controls-TreeGrid__node-header-expanderPadding {
+.controls-TreeGrid__node-extraItem-expanderPadding {
    line-height: 1;
    display: inline-block;
    position: relative;
