+++ tests/ControlsUnit/treeGrid_clean/Display/NodeFooter/NodeFooter.test.ts
@@ -73,8 +73,8 @@ describe('Controls/_display/itemsStrategy/NodeFooter', () => {
         strategy = new NodeFooter({
             display: tree,
             source,
-            nodeFooterVisibilityCallback: () => true,
-            nodeFooterModule: 'Controls/treeGrid:TreeGridNodeFooterRow'
+            extraItemVisibilityCallback: () => true,
+            extraItemModule: 'Controls/treeGrid:TreeGridNodeFooterRow'
         });
     });
 
@@ -111,29 +111,5 @@ describe('Controls/_display/itemsStrategy/NodeFooter', () => {
             const removedFooter = items.find((it) => it.getContents() === 'node-footer-2');
             assert.isNotOk(removedFooter);
         });
-    });
-
-    it('not call callback for node if show more button', () => {
-        let calledForNodeWithMore = false;
-
-        const nodeFooterVisibilityCallback = (contents) => {
-            calledForNodeWithMore = contents.getKey() === 1;
-            return true;
-        };
-
-        tree.setHasMoreStorage({
-            1: true
-        });
-
-        strategy = new NodeFooter({
-            display: tree,
-            source,
-            nodeFooterVisibilityCallback: () => true,
-            nodeFooterModule: 'Controls/treeGrid:TreeGridNodeFooterRow'
-        });
-        // tslint:disable-next-line:no-unused-expression
-        strategy.items;
-
-        assert.isFalse(calledForNodeWithMore);
     });
 });
